+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added `timelines_access` property to `/api/v2/instance` response ([#191](https://codeberg.org/silverpill/mitra/pulls/191)).
 - Added API endpoints for session management.
+- Allow adding subscribers when Monero integration is not enabled.
 
 ### Changed